+++ ipfsconn/ipfshttp/ipfshttp.go
@@ -10,18 +10,13 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
-	"net"
 	"net/http"
-	"net/http/httputil"
 	"net/url"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
 
-	"github.com/ipfs/ipfs-cluster/adder/adderutils"
 	"github.com/ipfs/ipfs-cluster/api"
-	"github.com/ipfs/ipfs-cluster/rpcutil"
 
 	rpc "github.com/hsanjuan/go-libp2p-gorpc"
 	cid "github.com/ipfs/go-cid"
@@ -44,14 +39,9 @@ var logger = logging.Logger("ipfshttp")
 var updateMetricMod = 10
 
 // Connector implements the IPFSConnector interface
-// and provides a component which does two tasks:
-//
-// On one side, it proxies HTTP requests to the configured IPFS
-// daemon. It is able to intercept these requests though, and
-// perform extra operations on them.
-//
-// On the other side, it is used to perform on-demand requests
-// against the configured IPFS daemom (such as a pin request).
+// and provides a component which  is used to perform
+// on-demand requests against the configured IPFS daemom
+// (such as a pin request).
 type Connector struct {
 	ctx    context.Context
 	cancel func()
@@ -59,14 +49,10 @@ type Connector struct {
 	config   *Config
 	nodeAddr string
 
-	handlers map[string]func(http.ResponseWriter, *http.Request)
-
 	rpcClient *rpc.Client
 	rpcReady  chan struct{}
 
-	listener net.Listener // proxy listener
-	server   *http.Server // proxy server
-	client   *http.Client // client to ipfs daemon
+	client *http.Client // client to ipfs daemon
 
 	updateMetricMutex sync.Mutex
 	updateMetricCount int
@@ -88,21 +74,9 @@ type ipfsPinLsResp struct {
 	Keys map[string]ipfsPinType
 }
 
-type ipfsPinOpResp struct {
-	Pins []string
-}
-
 type ipfsIDResp struct {
 	ID        string
 	Addresses []string
-}
-
-// From https://github.com/ipfs/go-ipfs/blob/master/core/coreunix/add.go#L49
-type ipfsAddResp struct {
-	Name  string
-	Hash  string `json:",omitempty"`
-	Bytes int64  `json:",omitempty"`
-	Size  string `json:",omitempty"`
 }
 
 type ipfsSwarmPeersResp struct {
@@ -142,38 +116,6 @@ func NewConnector(cfg *Config) (*Connector, error) {
 		return nil, err
 	}
 
-	proxyNet, proxyAddr, err := manet.DialArgs(cfg.ProxyAddr)
-	if err != nil {
-		return nil, err
-	}
-
-	l, err := net.Listen(proxyNet, proxyAddr)
-	if err != nil {
-		return nil, err
-	}
-
-	nodeHTTPAddr := "http://" + nodeAddr
-	proxyURL, err := url.Parse(nodeHTTPAddr)
-	if err != nil {
-		return nil, err
-	}
-
-	proxyHandler := httputil.NewSingleHostReverseProxy(proxyURL)
-
-	smux := http.NewServeMux()
-	s := &http.Server{
-		ReadTimeout:       cfg.ProxyReadTimeout,
-		WriteTimeout:      cfg.ProxyWriteTimeout,
-		ReadHeaderTimeout: cfg.ProxyReadHeaderTimeout,
-		IdleTimeout:       cfg.ProxyIdleTimeout,
-		Handler:           smux,
-	}
-
-	// See: https://github.com/ipfs/go-ipfs/issues/5168
-	// See: https://github.com/ipfs/ipfs-cluster/issues/548
-	// on why this is re-enabled.
-	s.SetKeepAlivesEnabled(false) // A reminder that this can be changed
-
 	c := &http.Client{} // timeouts are handled by context timeouts
 
 	ctx, cancel := context.WithCancel(context.Background())
@@ -183,22 +125,9 @@ func NewConnector(cfg *Config) (*Connector, error) {
 		config:   cfg,
 		cancel:   cancel,
 		nodeAddr: nodeAddr,
-		handlers: make(map[string]func(http.ResponseWriter, *http.Request)),
 		rpcReady: make(chan struct{}, 1),
-		listener: l,
-		server:   s,
 		client:   c,
 	}
-
-	smux.Handle("/", proxyHandler)
-	smux.HandleFunc("/api/v0/pin/add/", ipfs.pinHandler)
-	smux.HandleFunc("/api/v0/pin/rm/", ipfs.unpinHandler)
-	smux.HandleFunc("/api/v0/pin/ls", ipfs.pinLsHandler) // required to handle /pin/ls for all pins
-	smux.HandleFunc("/api/v0/pin/ls/", ipfs.pinLsHandler)
-	smux.HandleFunc("/api/v0/add", ipfs.addHandler)
-	smux.HandleFunc("/api/v0/add/", ipfs.addHandler)
-	smux.HandleFunc("/api/v0/repo/stat", ipfs.repoStatHandler)
-	smux.HandleFunc("/api/v0/repo/stat/", ipfs.repoStatHandler)
 
 	go ipfs.run()
 	return ipfs, nil
@@ -213,21 +142,6 @@ func (ipfs *Connector) run() {
 	// -- prevents race conditions with ipfs.wg.
 	ipfs.shutdownLock.Lock()
 	defer ipfs.shutdownLock.Unlock()
-
-	// This launches the proxy
-	ipfs.wg.Add(1)
-	go func() {
-		defer ipfs.wg.Done()
-		logger.Infof(
-			"IPFS Proxy: %s -> %s",
-			ipfs.config.ProxyAddr,
-			ipfs.config.NodeAddr,
-		)
-		err := ipfs.server.Serve(ipfs.listener) // hangs here
-		if err != nil && !strings.Contains(err.Error(), "closed network connection") {
-			logger.Error(err)
-		}
-	}()
 
 	// This runs ipfs swarm connect to the daemons of other cluster members
 	ipfs.wg.Add(1)
@@ -248,264 +162,6 @@ func (ipfs *Connector) run() {
 			return
 		}
 	}()
-}
-
-func ipfsErrorResponder(w http.ResponseWriter, errMsg string) {
-	res := ipfsError{errMsg}
-	resBytes, _ := json.Marshal(res)
-	w.Header().Add("Content-Type", "application/json")
-	w.WriteHeader(http.StatusInternalServerError)
-	w.Write(resBytes)
-	return
-}
-
-func (ipfs *Connector) pinOpHandler(op string, w http.ResponseWriter, r *http.Request) {
-	arg, ok := extractArgument(r.URL)
-	if !ok {
-		ipfsErrorResponder(w, "Error: bad argument")
-		return
-	}
-	c, err := cid.Decode(arg)
-	if err != nil {
-		ipfsErrorResponder(w, "Error parsing CID: "+err.Error())
-		return
-	}
-
-	err = ipfs.rpcClient.Call(
-		"",
-		"Cluster",
-		op,
-		api.PinCid(c).ToSerial(),
-		&struct{}{},
-	)
-	if err != nil {
-		ipfsErrorResponder(w, err.Error())
-		return
-	}
-
-	res := ipfsPinOpResp{
-		Pins: []string{arg},
-	}
-	resBytes, _ := json.Marshal(res)
-	w.Header().Add("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	w.Write(resBytes)
-	return
-}
-
-func (ipfs *Connector) pinHandler(w http.ResponseWriter, r *http.Request) {
-	ipfs.pinOpHandler("Pin", w, r)
-}
-
-func (ipfs *Connector) unpinHandler(w http.ResponseWriter, r *http.Request) {
-	ipfs.pinOpHandler("Unpin", w, r)
-}
-
-func (ipfs *Connector) pinLsHandler(w http.ResponseWriter, r *http.Request) {
-	pinLs := ipfsPinLsResp{}
-	pinLs.Keys = make(map[string]ipfsPinType)
-
-	arg, ok := extractArgument(r.URL)
-	if ok {
-		c, err := cid.Decode(arg)
-		if err != nil {
-			ipfsErrorResponder(w, err.Error())
-			return
-		}
-		var pin api.PinSerial
-		err = ipfs.rpcClient.Call(
-			"",
-			"Cluster",
-			"PinGet",
-			api.PinCid(c).ToSerial(),
-			&pin,
-		)
-		if err != nil {
-			ipfsErrorResponder(w, fmt.Sprintf("Error: path '%s' is not pinned", arg))
-			return
-		}
-		pinLs.Keys[pin.Cid] = ipfsPinType{
-			Type: "recursive",
-		}
-	} else {
-		var pins []api.PinSerial
-		err := ipfs.rpcClient.Call(
-			"",
-			"Cluster",
-			"Pins",
-			struct{}{},
-			&pins,
-		)
-		if err != nil {
-			ipfsErrorResponder(w, err.Error())
-			return
-		}
-
-		for _, pin := range pins {
-			pinLs.Keys[pin.Cid] = ipfsPinType{
-				Type: "recursive",
-			}
-		}
-	}
-
-	resBytes, _ := json.Marshal(pinLs)
-	w.Header().Add("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	w.Write(resBytes)
-}
-
-func (ipfs *Connector) addHandler(w http.ResponseWriter, r *http.Request) {
-	reader, err := r.MultipartReader()
-	if err != nil {
-		ipfsErrorResponder(w, "error reading request: "+err.Error())
-		return
-	}
-
-	q := r.URL.Query()
-	if q.Get("only-hash") == "true" {
-		ipfsErrorResponder(w, "only-hash is not supported when adding to cluster")
-	}
-
-	unpin := q.Get("pin") == "false"
-
-	// Luckily, most IPFS add query params are compatible with cluster's
-	// /add params. We can parse most of them directly from the query.
-	params, err := api.AddParamsFromQuery(q)
-	if err != nil {
-		ipfsErrorResponder(w, "error parsing options:"+err.Error())
-		return
-	}
-	trickle := q.Get("trickle")
-	if trickle == "true" {
-		params.Layout = "trickle"
-	}
-
-	logger.Warningf("Proxy/add does not support all IPFS params. Current options: %+v", params)
-
-	outputTransform := func(in *api.AddedOutput) interface{} {
-		r := &ipfsAddResp{
-			Name:  in.Name,
-			Hash:  in.Cid,
-			Bytes: int64(in.Bytes),
-		}
-		if in.Size != 0 {
-			r.Size = strconv.FormatUint(in.Size, 10)
-		}
-		return r
-	}
-
-	root, err := adderutils.AddMultipartHTTPHandler(
-		ipfs.ctx,
-		ipfs.rpcClient,
-		params,
-		reader,
-		w,
-		outputTransform,
-	)
-
-	// any errors have been sent as Trailer
-	if err != nil {
-		return
-	}
-
-	if !unpin {
-		return
-	}
-
-	// Unpin because the user doesn't want to pin
-	time.Sleep(100 * time.Millisecond)
-	err = ipfs.rpcClient.CallContext(
-		ipfs.ctx,
-		"",
-		"Cluster",
-		"Unpin",
-		api.PinCid(root).ToSerial(),
-		&struct{}{},
-	)
-	if err != nil {
-		w.Header().Set("X-Stream-Error", err.Error())
-		return
-	}
-}
-
-func (ipfs *Connector) repoStatHandler(w http.ResponseWriter, r *http.Request) {
-	var peers []peer.ID
-	err := ipfs.rpcClient.Call(
-		"",
-		"Cluster",
-		"ConsensusPeers",
-		struct{}{},
-		&peers,
-	)
-	if err != nil {
-		ipfsErrorResponder(w, err.Error())
-		return
-	}
-
-	ctxs, cancels := rpcutil.CtxsWithTimeout(ipfs.ctx, len(peers), ipfs.config.IPFSRequestTimeout)
-	defer rpcutil.MultiCancel(cancels)
-
-	repoStats := make([]api.IPFSRepoStat, len(peers), len(peers))
-	repoStatsIfaces := make([]interface{}, len(repoStats), len(repoStats))
-	for i := range repoStats {
-		repoStatsIfaces[i] = &repoStats[i]
-	}
-
-	errs := ipfs.rpcClient.MultiCall(
-		ctxs,
-		peers,
-		"Cluster",
-		"IPFSRepoStat",
-		struct{}{},
-		repoStatsIfaces,
-	)
-
-	totalStats := api.IPFSRepoStat{}
-
-	for i, err := range errs {
-		if err != nil {
-			logger.Errorf("%s repo/stat errored: %s", peers[i], err)
-			continue
-		}
-		totalStats.RepoSize += repoStats[i].RepoSize
-		totalStats.StorageMax += repoStats[i].StorageMax
-	}
-
-	resBytes, _ := json.Marshal(totalStats)
-	w.Header().Add("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	w.Write(resBytes)
-	return
-}
-
-// SetClient makes the component ready to perform RPC
-// requests.
-func (ipfs *Connector) SetClient(c *rpc.Client) {
-	ipfs.rpcClient = c
-	ipfs.rpcReady <- struct{}{}
-}
-
-// Shutdown stops any listeners and stops the component from taking
-// any requests.
-func (ipfs *Connector) Shutdown() error {
-	ipfs.shutdownLock.Lock()
-	defer ipfs.shutdownLock.Unlock()
-
-	if ipfs.shutdown {
-		logger.Debug("already shutdown")
-		return nil
-	}
-
-	logger.Info("stopping IPFS Proxy")
-
-	ipfs.cancel()
-	close(ipfs.rpcReady)
-	ipfs.server.SetKeepAlivesEnabled(false)
-	ipfs.listener.Close()
-
-	ipfs.wg.Wait()
-	ipfs.shutdown = true
-	return nil
 }
 
 // ID performs an ID request against the configured
